-Additional longs to int changes. -Regenrated scaffolding from scratch. The db update was failing because of some FK BS. Easier this way for now.

+++ src/Company.Videomatic.Application/Features/Artifacts/Commands/CreateArtifact.cs
@@ -1,6 +1,6 @@
 ﻿namespace Company.Videomatic.Application.Features.Artifacts.Commands;
 
-public record CreateArtifactCommand(long VideoId,
+public record CreateArtifactCommand(int VideoId,
                                     string Name,
                                     string Type,
                                     string? Text) : CreateAggregateRootCommand<Artifact>();
